Address review: move WorldList to util.list, drop entry trimming
